python/src/auth/requirements.txt
astroid==4.0.3
blinker==1.9.0
click==8.3.1
dill==0.4.1
Flask==3.1.2
Flask-MySQLdb==2.0.0
isort==7.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mysqlclient==2.2.7
parso==0.8.5
platformdirs==4.5.1
PyJWT==2.11.0
pylint==4.0.4
python-dotenv==1.2.1
tomlkit==0.14.0
Werkzeug==3.1.5
